requirements.txt
demoparser2
python-dotenv
requests